requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Energy Modelling
pypsa==0.26.1
numpy==1.26.4
pandas==2.2.2
numpy-financial==1.0.0

# Scientific / Numerical
scipy==1.13.0

# Financial
numpy-financial==1.0.0

# Visualisation
plotly==5.22.0
matplotlib==3.8.4

# Reporting
jinja2==3.1.4

# Caching
diskcache==5.6.3

# Frontend
streamlit==1.33.0
streamlit-authenticator==0.3.1

# Testing
pytest==8.2.0
httpx==0.27.0
